requirements.txt
asgiref==3.2.10 
django==3.1.2 
pytz==2020.1 
sqlparse==0.4.1
djangorestframework==3.12.1
numpy==1.19.2
certifi==2020.6.20 
cycler==0.10.0 
kiwisolver==1.2.0 
matplotlib==3.3.2 
pillow==8.0.0 
pyparsing==2.4.7 
python==dateutil==2.8.1 
six==1.15.0
opencv-python==4.4.0.44